+++ vitest.config.ts
@@ -2,6 +2,6 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: ['utils/**/*.test.ts'],
+    include: ['utils/**/*.test.ts', 'components/**/*.test.ts'],
   },
 });